Update shinyTimer widget

Update shinyTimer widget

Usage

updateShinyTimer(
  inputId,
  hours = NULL,
  minutes = NULL,
  seconds = NULL,
  type = NULL,
  label = NULL,
  frame = NULL,
  session = shiny::getDefaultReactiveDomain()
)

Arguments

inputId

The input ID corresponding to the UI element.

hours

The new starting time in hours for the countdown.

minutes

The new starting time in minutes for the countdown.

seconds

The new starting time in seconds for the countdown.

type

The new type of the countdown timer display ("simple", "mm:ss", "hh:mm:ss", "mm:ss.cs").

label

The new label to be displayed above the countdown timer.

frame

The new shape of the timer's container ("none", "circle", "rectangle").

session

The session object from the shiny server function.

Value

No return value, called for side effects.

Examples

if (interactive()) {
  library(shiny)
  shinyApp(
    ui = fluidPage(
      shinyTimer("timer", label = "Countdown Timer", seconds = 10, type = "mm:ss"),
      actionButton("update", "Update Timer")
    ),
    server = function(input, output, session) {
      observeEvent(input$update, {
        updateShinyTimer("timer", seconds = 20, type = "hh:mm:ss")
      })
    }
  )
}